requirements.txt
feedparser
pyyaml
markdownify
requests
slack-sdk